pyproject.toml
[project]
name = "abstraction-agent"
version = "0.1.0"
description = "A robust LangGraph-based agentic framework with native GUI"
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "torch>=2.0.0",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
    "bitsandbytes>=0.43.0",
    "PySide6>=6.6.0",
    "PySide6-Addons>=6.6.0",
    "diskcache>=5.6.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "black>=23.0.0",
    "pylint>=3.0.0",
    "autopep8>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "mypy>=1.7.0",
    "huggingface-hub>=0.23.0",
    "mcp>=1.0.0",
    "anyio>=4.0.0",
    "starlette>=0.27.0",
]

[project.scripts]
abstraction-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# PyApp Configuration
[tool.pyapp]
# Embed Python interpreter
python-version = "3.11"

# Distribution method
distribution-method = "pip"

# Entry point
entry-point = "src.main:main"

# GUI mode (no console window on Windows)
gui = true

# Include all dependencies
install-dependencies = true
